Move LocalDestination instantiation to its own trait
