+++ api/src/main/scala/quasar/api/datasource/DestinationType.scala
@@ -28,7 +28,7 @@ import scalaz.std.tuple._
 import scalaz.syntax.show._
 
 @Lenses
-final case class DestinationType(name: DestinationType.Name, version: Positive)
+final case class DestinationType(name: DestinationType.Name, version: Positive, minSupportedVersion: Positive)
 
 object DestinationType extends DestinationTypeInstances {
   type Name = datasource.Name
