+++ connector/src/main/scala/quasar/connector/DestinationModule.scala
@@ -20,15 +20,13 @@ import quasar.api.datasource.DatasourceError.InitializationError
 import quasar.api.datasource.DestinationType
 
 import argonaut.Json
-import cats.effect.{Effect, ContextShift, Resource}
+import cats.effect.{ConcurrentEffect, ContextShift, Resource, Timer}
 import scalaz.\/
 
 trait DestinationModule {
-  def destinationKind: DestinationType
+  def destinationType: DestinationType
 
-  def sanitizeDestinationConfig(config: Json): Json
-
-  def destination[F[_]: Effect: ContextShift: MonadResourceErr](
+  def destination[F[_]: ConcurrentEffect: ContextShift: MonadResourceErr: Timer](
     config: Json)
     : F[InitializationError[Json] \/ Resource[F, Destination[F]]]
 }
